+++ kafka/integration_test.go
@@ -1377,7 +1377,7 @@ func (its *IntegrationTestSuite) TestAdminConfig() {
 	}
 }
 
-//Test AdminClient GetMetadata API
+// Test AdminClient GetMetadata API
 func (its *IntegrationTestSuite) TestAdminGetMetadata() {
 	t := its.T()
 
@@ -1659,7 +1659,131 @@ func (its *IntegrationTestSuite) TestAdminACLs() {
 	checkExpectedResult(expectedDescribeACLs, *resultDescribeACLs)
 }
 
-//Test consumer QueryWatermarkOffsets API
+// Test AdminClient List all consumer group offsets.
+func (its *IntegrationTestSuite) TestAdminClient_ListAllConsumerGroupsOffsets() {
+	t := its.T()
+	rand.Seed(time.Now().Unix())
+
+	conf := &ConfigMap{"bootstrap.servers": testconf.Brokers}
+	if err := conf.updateFromTestconf(); err != nil {
+		t.Fatalf("Failed to update test configuration: %s\n", err)
+	}
+
+	admin, err := NewAdminClient(conf)
+	if err != nil {
+		t.Fatalf("Failed to create Admin client: %s\n", err)
+	}
+	defer admin.Close()
+
+	// Create some topics.
+	numTopics := 3
+	topics := make([]string, 0)
+	topicSpec := make([]TopicSpecification, 0)
+
+	for i := 0; i < numTopics; i++ {
+		topic := fmt.Sprintf("%s-%d", testconf.Topic, rand.Intn(100000))
+		topics = append(topics, topic)
+		topicSpec = append(
+			topicSpec, TopicSpecification{Topic: topic, NumPartitions: i + 1})
+	}
+
+	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
+	defer cancel()
+	topicResult, err := admin.CreateTopics(ctx, topicSpec)
+
+	if err != nil {
+		t.Fatalf("Failed to create topics: %s\n", err)
+	}
+
+	for i := 0; i < numTopics; i++ {
+		if topicResult[i].Error.Code() != ErrNoError {
+			t.Fatalf("Failed to create topic %s: %s", topicSpec[i].Topic, topicResult[i].Error)
+		}
+	}
+
+	// Join a consumer group and subscribe to the created topics,
+	// commit some offsets to read later.
+	group := fmt.Sprintf("%s-%d", testconf.GroupID, rand.Intn(100000))
+	conf = &ConfigMap{
+		"bootstrap.servers": testconf.Brokers,
+		"group.id":          group,
+		"auto.offset.reset": "end",
+	}
+	conf.updateFromTestconf()
+
+	consumer, err := NewConsumer(conf)
+	if err != nil {
+		t.Fatalf("Failed to create consumer: %s\n", err)
+	}
+	defer consumer.Close()
+
+	if err = consumer.SubscribeTopics(topics, nil); err != nil {
+		t.Fatalf("Failed to subscribe to topics: %s\n", err)
+	}
+
+	// Poll for a while, wait for rebalance.
+	consumer.Poll(10000)
+
+	for i := 0; i < numTopics; i++ {
+		for j := 0; j < topicSpec[i].NumPartitions; j++ {
+			_, err = consumer.CommitOffsets([]TopicPartition{
+				{
+					Topic:     &topics[i],
+					Partition: int32(j),
+					Offset:    Offset((i * topicSpec[i].NumPartitions) + j),
+				},
+			})
+			if err != nil {
+				t.Fatalf("Could not commit for %s:%d: %s\n", topics[i], j, err)
+			}
+		}
+	}
+
+	// List consumer group offsets.
+	ctx, cancel = context.WithTimeout(context.Background(), 10*time.Second)
+	defer cancel()
+	offsets, err := admin.ListConsumerGroupOffsets(
+		ctx,
+		[]ConsumerGroupTopicPartitions{{Group: group}},
+		SetAdminRequireStableOffsets(true))
+
+	if err != nil {
+		t.Fatalf("Failed to get consumer group offsets: %s\n", err)
+	}
+
+	if len(offsets.ConsumerGroupsTopicPartitions) != 1 {
+		t.Fatal("Consumer group offsets is empty")
+	}
+
+	toppars := offsets.ConsumerGroupsTopicPartitions[0].Partitions
+
+	// Use linear search - okay for small numTopics. Since the returned list
+	// is not ordered in any particular way.
+	matchedToppars := 0
+	for i := 0; i < numTopics; i++ {
+		for k := 0; k < len(toppars); k++ {
+			if topics[i] != *toppars[k].Topic {
+				continue
+			}
+			matchedToppars++
+			partition := int(toppars[k].Partition)
+			expectedOffset := i*topicSpec[i].NumPartitions + partition
+
+			if toppars[k].Offset != Offset(expectedOffset) {
+				t.Fatalf("Expected offset %d for %s:%d, got %d\n",
+					expectedOffset, topics[i], partition, toppars[k].Offset)
+			}
+		}
+	}
+
+	totalPartitions := ((numTopics + 1) * numTopics) / 2
+	if matchedToppars != totalPartitions {
+		t.Fatalf("Expected to match %d total topic partitions, matched %d\n",
+			totalPartitions, matchedToppars)
+	}
+}
+
+// Test consumer QueryWatermarkOffsets API
 func (its *IntegrationTestSuite) TestConsumerQueryWatermarkOffsets() {
 	t := its.T()
 
@@ -1690,7 +1814,7 @@ func (its *IntegrationTestSuite) TestConsumerQueryWatermarkOffsets() {
 
 }
 
-//Test consumer GetWatermarkOffsets API
+// Test consumer GetWatermarkOffsets API
 func (its *IntegrationTestSuite) TestConsumerGetWatermarkOffsets() {
 	t := its.T()
 
@@ -1745,7 +1869,7 @@ func (its *IntegrationTestSuite) TestConsumerGetWatermarkOffsets() {
 
 }
 
-//TestConsumerOffsetsForTimes
+// TestConsumerOffsetsForTimes
 func (its *IntegrationTestSuite) TestConsumerOffsetsForTimes() {
 	t := its.T()
 
@@ -1840,7 +1964,7 @@ func (its *IntegrationTestSuite) TestConsumerGetMetadata() {
 	t.Logf("Meta data for consumer: %v\n", metaData)
 }
 
-//Test producer QueryWatermarkOffsets API
+// Test producer QueryWatermarkOffsets API
 func (its *IntegrationTestSuite) TestProducerQueryWatermarkOffsets() {
 	t := its.T()
 
@@ -1882,7 +2006,7 @@ func (its *IntegrationTestSuite) TestProducerQueryWatermarkOffsets() {
 	}
 }
 
-//Test producer GetMetadata API
+// Test producer GetMetadata API
 func (its *IntegrationTestSuite) TestProducerGetMetadata() {
 	t := its.T()